refactor(export): extract the property/quantity-set phases into step-property-sets.ts (#2475 steps 2b+2c)
Moves `export()`'s property-set and quantity-set collection loops (`:747-1000`),
its three generation loops (`:1193-1361` — pset generation, the type-object
`HasPropertySets` rewrite, qset generation) and the fifteen helpers they own out
of `step-exporter.ts`, which drops 2392 -> 1542 lines. `step-exporter.ts` keeps
two call sites.

2b and 2c are one step because 2c is not separable: `generatedTypeOwnedPsetIds`
is written by the pset generation loop and read by the rewrite loop, and
`ExportPass` documents it as the one binding "read in one phase only". Putting
all three loops in `generatePropertyAndQuantitySetEntities` keeps it a plain
local — not a pass field, not an argument threaded between two functions. The
same coupling runs the other way: `getTypeOwnedHasPropertySetIds` is 2c's helper
but its only call site is 2b's collection loop, and `getPropertySetName` is
called from both.

The phases are not pure functions of their inputs: they allocate express ids
from the exporter's `nextExpressId` (all six sites in the export live in the two
generators), read the store through its `EntityExtractor` and `MutablePropertyView`,
and memoise owner-history lookups. Those are injected as a `PropertySetContext`,
following `step-georeferencing.ts:GeorefContext` from step 2a, rather than
hoisted onto the pass. The owner-history memos stay OWNED by the exporter and
are passed by reference: their reset is an `export()`-level statement whose
comment is where "per export, not per exporter" is argued, and storing them on a
per-export context would make that reset implicit.

Three symbols had a consumer outside the moved region and are exported rather
than module-private, the shape step 2a used for `findLengthUnitReference`:
`getPropertyIdsInSet` (read by `retainSharedAtoms`, which belongs to neither
phase and runs between them), `buildRelDefinesByPropertiesIndex` (its
`relatedByRel` half feeds `export()`'s deleted-host relationship sweep), and
`entityLineText` / `getPropertySetName` (the pin in
`entity-line-text-bounds.test.ts`, which used to reach them by casting a
`StepExporter` to an interface of its privates). `decodeRange` moves to
`source-ref-bounds.ts` so both files share one decode; `applySourceLineMutations`
is shared with the source-iteration pass and is injected, not moved.

Verified per the standing per-site rule: all 32 guards and ledger/warning call
sites in the moved regions were mutated individually before and after the move,
and every one killed a byte-identical set of test NAMES. Suite 738 tests / 708
passed / 30 skipped before and after, with identical full-name sets.

Test membership was established by disabling each region and recording what
dies (55 failures / 17 files for the generators, 74 / 16 for the collection, 34
/ 8 for the rewrite — reproducing the scoping measurement exactly), not by name.
Only two of those tests needed relocating: they are the whole of what dies in
`step-exporter.test.ts`, and they move to `step-property-sets.test.ts` with
their `describe`/`it` names unchanged. Everything else stays where it is —
every test file in this package lives in one flat directory, so a file that is
already the dedicated test of a unit that moved (`owner-history`,
`pset-nomination-effect`, `quantity-set-withholding`,
`rewritten-type-line-attributes`, `type-object-rewrite-fallback`,
`slot5-caller-audit`, `entity-line-text-bounds`) already names its code, and a
file that belongs to a unit that did NOT move (the delta ledger, the overlay
effective model, the declared property type, `retainSharedAtoms`) should not
follow psets out.

The unused-locals baseline tightens by one: `type InPlaceNominees` was imported
into `step-exporter.ts` and referenced only from a comment.

No behaviour change, so no changeset.
